some console refactoring, display network results in ui
