+++ frontend/app/components/Spots/SpotPlayer/spotPlayerStore.ts
@@ -56,7 +56,7 @@ const mapSpotNetworkToEv = (ev: SpotNetworkRequest): any => {
   })
   const response = JSON.stringify({
     headers: ev.responseHeaders,
-    body: { warn: "Chrome Manifest V3 -- No response body available in Chrome 93+" }
+    body: ev.responseBody ?? { warn: "Chrome Manifest V3 -- No response body available in Chrome 93+" }
   })
   return ({
     ...ev,
